+++ tests/test_subset.py
@@ -1979,3 +1979,76 @@ def test_passed_coords(data_dir, subset_output_dir):
     assert lats == dummy_lats
     assert lons == dummy_lons
     assert times == dummy_times
+
+
+def test_var_subsetting_tropomi(data_dir, subset_output_dir, request):
+    """
+    Check that variable subsetting is the same if a leading slash is included
+    """
+    trop_dir = join(data_dir, 'tropomi')
+    trop_file = 'S5P_OFFL_L2__CH4____20190319T110835_20190319T125006_07407_01_010202_20190325T125810_subset.nc4'
+    variable_slash = ['/PRODUCT/methane_mixing_ratio']
+    variable_noslash = ['PRODUCT/methane_mixing_ratio']
+    bbox = np.array(((-180, 180), (-90, 90)))
+    output_file_slash = "{}_{}".format(request.node.name, trop_file)
+    output_file_noslash = "{}_noslash_{}".format(request.node.name, trop_file)
+    shutil.copyfile(
+        os.path.join(trop_dir, trop_file),
+        os.path.join(subset_output_dir, trop_file)
+    )
+    shutil.copyfile(
+        os.path.join(trop_dir, trop_file),
+        os.path.join(subset_output_dir, 'slashtest' + trop_file)
+    )
+    subset.subset(
+        file_to_subset=join(subset_output_dir, trop_file),
+        bbox=bbox,
+        output_file=join(subset_output_dir, output_file_slash),
+        variables=variable_slash
+    )
+    subset.subset(
+        file_to_subset=join(subset_output_dir, 'slashtest' + trop_file),
+        bbox=bbox,
+        output_file=join(subset_output_dir, output_file_noslash),
+        variables=variable_noslash
+    )
+
+    slash_dataset = nc.Dataset(join(subset_output_dir, output_file_slash))
+    noslash_dataset = nc.Dataset(join(subset_output_dir, output_file_noslash))
+
+    assert list(slash_dataset.groups['PRODUCT'].variables) == list(noslash_dataset.groups['PRODUCT'].variables)
+
+
+def test_bad_time_unit(subset_output_dir):
+
+    fill_val = -99999.0
+    time_vals = np.random.rand(10)
+    time_vals[0] = fill_val
+    time_vals[-1] = fill_val
+
+    data_vars = {
+        'foo': (['x'], np.random.rand(10)),
+        'time': (
+            ['x'],
+            time_vals,
+            {
+                'units': 'seconds since 2000-1-1 0:0:0 0',
+                '_FillValue': fill_val,
+                'standard_name': 'time',
+                'calendar': 'standard'
+            }
+        ),
+    }
+
+    ds = xr.Dataset(
+        data_vars=data_vars,
+        coords={'x': (['x'], np.arange(10))}
+    )
+
+    nc_out_location = join(subset_output_dir, "bad_time.nc")
+    ds.to_netcdf(nc_out_location)
+
+    subset.override_decode_cf_datetime()
+
+    ds_test = xr.open_dataset(nc_out_location)
+    ds_test.close()